Move ClusterServiceStatus to types.ts
